Hoan thanh giai ngan

+++ DTO/GN_KH.cs
@@ -11,7 +11,7 @@ namespace DTO
         private string soTKLK;
         private string tenKH;
         private string maGN;
-        private int soTienGN;
+        private long soTienGN;
         private string trangThai;
         private int duNoGoc;
         private int duNoLaiTrongHan;
@@ -20,7 +20,7 @@ namespace DTO
         public string SoTKLK { get; set; }
         public string TenKH { get; set; }
         public string MaGN { get; set; }
-        public int SoTienGN { get; set; }
+        public long SoTienGN { get; set; }
         public string TrangThai { get; set; }
         public int DuNoGoc { get; set; }
         public int DuNoLaiTrongHan { get; set; }
